+++ oc/compiler/static_analyzer/static_analyzer.c
@@ -5,13 +5,68 @@
  */
 
 #include "static_analyzer.h"
+#include <sys/types.h>
 
 //Store these globally for easy access
 static three_addr_var_t* instruction_pointer_var;
 static three_addr_var_t* stack_pointer_var;
 
+//Hold references to the error/warning counts
+static u_int32_t* error_count;
+static u_int32_t* warning_count;
+
+//The current dependency node(changed by function)
+static dependency_graph_node_t* current_dependency_node;
+
+//Holder for all error info that we want to print
+static char error_info[ERROR_SIZE * 3];
 
 //========================================= General Utilities =============================================
+
+/**
+ * Take a file that may look like: ./oc/test_files/sample.ol and return sample.ol
+ */
+static inline char* extract_file_name_from_fully_qualified_name(char* fully_qualified_name){
+	int32_t length = strlen(fully_qualified_name);
+
+	//Roll this back until we have the index of the first /
+	int32_t i = length - 1;
+	for(; i >= 0; i--){
+		if(fully_qualified_name[i] == '/'){
+			break;
+		}
+	}
+
+	//Offset into this to get it(+ 1 to get past the /)
+	return fully_qualified_name + i + 1;
+}
+
+
+/**
+ * Simply prints a parse message in a nice formatted way. For the CFG, there
+ * are no parser line numbers
+ *
+ *
+ * TODO REVAMP THIS
+*/
+static void print_cfg_message(error_message_type_t message_type, char* info, u_int32_t line_number){
+	//Mapped by index to the enum values
+	static const char* type[] = {"WARNING", "ERROR", "INFO", "DEBUG"};
+
+	//Get just the important part of the file name
+	char* file_name = extract_file_name_from_fully_qualified_name(current_dependency_node->file_name);
+
+	/**
+	 * If it's the main node print out the file only, otherwise we'll
+	 * also need the module name
+	 */
+	if(current_dependency_node->type != DEPENDENCY_GRAPH_NODE_TYPE_MAIN){
+		fprintf(stdout, "\n[MODULE %s | FILE: %s] --> [LINE %d | COMPILER %s]: %s\n", current_dependency_node->module_name.string, file_name, line_number, type[message_type], info);
+	} else {
+		fprintf(stdout, "\n[FILE: %s] --> [LINE %d | COMPILER %s]: %s\n", file_name, line_number, type[message_type], info);
+	}
+}
+
 /**
  * Run through an entire array of function blocks and reset the status for
  * every single one. We assume that the caller knows what they are doing, and
@@ -66,6 +121,25 @@ static inline u_int8_t does_variable_dynamic_array_contain_symtab_variable(dynam
 		}
 	}
 
+	return FALSE;
+}
+
+
+/**
+ * This simple utility will scan a dynamic array of variables and invoke the variables_equal() function
+ * on each of them for the given variable
+ */
+static inline u_int8_t does_variable_dynamic_array_contain_variable(dynamic_array_t* array, three_addr_var_t* variable){
+	for(int32_t i = 0; i < array->current_index; i++){
+		three_addr_var_t* candidate = dynamic_array_get_at(array, i);
+
+		//If we have one equals then the whole thing works
+		if(variables_equal(candidate, variable) == TRUE){
+			return TRUE;
+		}
+	}
+
+	//If we made it here then no match
 	return FALSE;
 }
 
@@ -858,6 +932,287 @@ static void convert_cfg_to_ssa_form(cfg_t* cfg, variable_symtab_t* variables){
 	rename_all_variables(cfg);
 }
 
+/**
+ * TODO
+ */
+static u_int8_t check_variable_for_definite_assignment(instruction_t* instruction, three_addr_var_t* variable, dynamic_array_t* may_not_have_been_initialized){
+	/**
+	 * First guard - if the variable is NULL(common) or it's not 
+	 * SSA eligible(temp var, etc.) we can leave now. This is still
+	 * a SUCCESS because there's nothing wrong with this
+	 */
+	if(variable == NULL || is_variable_ssa_eligible(variable) == FALSE){
+		return SUCCESS;
+	}
+
+	/**
+	 * The stack and instruction pointer are special cases that are exempt from this
+	 * kind of checking so leave if we see it
+	 */
+	if(variable == stack_pointer_var || variable == instruction_pointer_var){
+		return SUCCESS;
+	}
+
+	/**
+	 * Obvious case - generation of 0 means it's never
+	 * been initialized so this is a pure use before
+	 * initialization
+	 */
+	if(variable->ssa_generation == 0){
+		sprintf(error_info, "Variable %s is used before initialization", variable->linked_var->var_name.string);
+		print_cfg_message(MESSAGE_TYPE_ERROR, error_info, instruction->line_number);
+		(*error_count)++;
+		return FAILURE;
+
+	/**
+	 * Not so obvious case - it being in this array means that it comes from a phi function
+	 * that itself had a parameter of generation 0, meaning that it's not a guarantee that
+	 * this wasn't initialized but it may not have been, which is still an error
+	 */
+	} else if(does_variable_dynamic_array_contain_variable(may_not_have_been_initialized, variable) == TRUE){
+		sprintf(error_info, "Variable %s may be used before initialization", variable->linked_var->var_name.string);
+		print_cfg_message(MESSAGE_TYPE_ERROR, error_info, instruction->line_number);
+		(*error_count)++;
+		return FAILURE;
+
+	/**
+	 * If we made it here then this worked and the variable is clean
+	 */
+	} else {
+		return SUCCESS;
+	}
+}
+
+
+/**
+ *
+ * TODO
+ */
+static inline u_int8_t does_instruction_comply_with_definite_assignment(instruction_t* instruction, dynamic_array_t* may_not_have_been_initialized){
+	//By default assume SUCCESS(1)
+	u_int8_t overall_result = SUCCESS;
+
+	/**
+	 * Regular non-phi function handling involves us checking every
+	 * single variable to see if we have any "_0" variables in use.
+	 * "_0" is our canary SSA value that represents an uninitialzed
+	 * variable. We will also need to make sure that each variable
+	 * is not a member of the "may_not_have_been_initialized" array.
+	 * This gets built up from phi functions who have values that may
+	 * have never been initialized
+	 *
+	 * We bitwise and the results together for this. One false in the chain
+	 * will make the whole thing 0(FAILURE)
+	 */
+	if(instruction->statement_type != THREE_ADDR_CODE_PHI_FUNC){
+		overall_result &= check_variable_for_definite_assignment(instruction, instruction->operands.oir.operand1, may_not_have_been_initialized);
+		overall_result &= check_variable_for_definite_assignment(instruction, instruction->operands.oir.operand2, may_not_have_been_initialized);
+		overall_result &= check_variable_for_definite_assignment(instruction, instruction->operands.oir.address_operand1, may_not_have_been_initialized);
+		overall_result &= check_variable_for_definite_assignment(instruction, instruction->operands.oir.address_operand2, may_not_have_been_initialized);
+
+		//Check all parameters as well
+		for(int32_t i = 0; i < instruction->parameters.current_index; i++){
+			three_addr_var_t* parameter = dynamic_array_get_at(&(instruction->parameters), i);
+
+			overall_result &= check_variable_for_definite_assignment(instruction, parameter, may_not_have_been_initialized);
+		}
+
+
+	/**
+	 * Phi-functions have special handling. If we have a phi
+	 * function that has at least one _0 variable in it, then the
+	 * LHS value may not have been initialized
+	 *
+	 * x_3 <- phi(x_2, x_1, x_0)
+	 * x_4 <- x_3 + 1
+	 *
+	 * Fail there, x_3 may not have been initialized. We will maintain
+	 * a list of variables that may be uninitialized that will be cross
+	 * referenced by all other checks. The value x_3 in this case would go
+	 * into there
+	 */
+	} else {
+		/**
+		 * Run through all of the parameters - all it takes is for one
+		 *
+		 * declare x:mut i32;
+		 *
+		 * if(<cond>){
+		 * 		x = 3;
+		 * } 
+		 *
+		 * ret x; <--- x may be used uninitialized here
+		 *
+		 */
+		for(int32_t i = 0; i < instruction->parameters.current_index; i++){
+			three_addr_var_t* parameter = dynamic_array_get_at(&(instruction->parameters), i);
+
+			/**
+			 * If we see a parameter with a generation of 0, that means that it's 
+			 * never been initialized. We will add the LHS to the "may_not_have_been_initialized"
+			 * list for later checks
+			 */
+			if(parameter->ssa_generation == 0){
+				dynamic_array_add(may_not_have_been_initialized, instruction->operands.oir.assignee);
+				overall_result = FAILURE;
+				
+				//We don't need to check any further for this
+				break;
+
+			/**
+			 * Just because it's not 0 doesn't mean we're safe. We must also check if this value
+			 * is inside of the "may not be initialized" list. If it is, then subsequently the
+			 * LHS of this instruction is also potentially uninitialized
+			 */
+			} else if(does_variable_dynamic_array_contain_variable(may_not_have_been_initialized, parameter) == TRUE){
+				dynamic_array_add(may_not_have_been_initialized, instruction->operands.oir.assignee);
+				overall_result = FAILURE;
+				
+				//We don't need to check any further for this
+				break;
+			}
+		}
+	}
+
+	return overall_result;
+}
+
+
+
+/**
+ * TODO
+ */
+static u_int8_t perform_definite_assignment_analysis_for_block(basic_block_t* block, dynamic_array_t* may_not_have_been_initialized){
+	//Assume success off the bat
+	u_int8_t result = SUCCESS;
+
+	//Grab a leader statement out
+	instruction_t* cursor = block->leader_statement;
+
+	/**
+	 * Go through every single statement and analyze every
+	 * variable within. If any of the statements is using
+	 * a _0 version, that counts as a use-before-initialize
+	 * and it is an error. If we have a phi-function that
+	 * has a _0 parameter, that is "maybe" use before initialize
+	 * and it is still an error
+	 *
+	 * NOTE: we do *NOT* check anything to do with the so-called "rip_offset_var". This 
+	 * is not a variable in the true sense so it's not worth it to mess around with it
+	 */
+	while(cursor != NULL){
+		/**
+		 * Any one instruction failing definite assignment means that the whole
+		 * thing fails. We will process all instructions to get a full picture of the
+		 * errors though
+		 */
+		if(does_instruction_comply_with_definite_assignment(cursor, may_not_have_been_initialized) == FALSE){
+			result = FALSE;
+		}
+
+		cursor = cursor->next_statement;
+	}
+
+	/**
+	 * For all dominator children of this block, go through and perform the
+	 * definite assignment analysis
+	 */
+	for(int32_t i = 0; i < block->dominator_children.current_index; i++){
+		basic_block_t* child = dynamic_array_get_at(&(block->dominator_children), i);
+
+		/**
+		 * If anything in this child fails, our overall result is failure. We will
+		 * keep going to scan everything though
+		 */
+		if(perform_definite_assignment_analysis_for_block(child, may_not_have_been_initialized) == FALSE){
+			result = FAILURE;
+		}
+	}
+
+	return result;
+}
+
+
+/**
+ * Perform the Ollie analyzer's version of definite assignment analysis. 
+ *
+ * This will include:
+ * 	1.) Uninitialized/potentially uninitialized variable detection
+ * 	2.) Unneccessary mutability detection
+ *
+ * TODO DOC
+ *
+ * We will scan all functions at once. If one function fails, we will still keep going to analyze
+ * the rest of the program. However, one function failing does mean that the entire program fails
+ * to compile in the end
+ */
+static inline u_int8_t perform_definite_assignment_analysis(cfg_t* cfg, variable_symtab_t* variables){
+	//Assume success off the bat
+	u_int8_t result = SUCCESS;
+
+	/**
+	 * Keep an array of variables that may not have been initialized in each function to make
+	 * scanning easier and less intensive. We'll allocate once and just wipe every time
+	 */
+	dynamic_array_t may_not_have_been_initialized = dynamic_array_alloc();
+
+	//Run through all functions
+	for(int32_t i = 0; i < cfg->function_entry_blocks.current_index; i++){
+		//Use the function entry to seed the search
+		basic_block_t* function_entry = dynamic_array_get_at(&(cfg->function_entry_blocks), i);
+
+		/**
+		 * Update the dependency node so that we get accurate error printouts
+		 */
+		current_dependency_node = function_entry->function_defined_in->dependency_graph_node;
+
+		/**
+		 * Call into the recursive analyzer. If we have a failure, then the entire thing
+		 * goes into failure, but we will keep scanning to get all errors in at once
+		 */
+
+		//TODO WE MAY ACTUALLY WANT TO STOP FAILURES AT JUST ONE HERE
+		if(perform_definite_assignment_analysis_for_block(function_entry, &may_not_have_been_initialized) == FAILURE){
+			result = FAILURE;
+		}
+
+		//Clear it now that we're done with this function
+		clear_dynamic_array(&may_not_have_been_initialized);
+	}
+
+	//Done with this so scrap it now
+	dynamic_array_dealloc(&may_not_have_been_initialized);
+
+	return result;
+}
+
+
+/**
+ * Perform mutability checking on the variable symtab. This will check
+ * and display warnings if we are labeling a variable as mutable
+ * but then never mutating it
+ */
+static void perform_mutability_checking(cfg_t* cfg, variable_symtab_t* symtab){
+
+	for(int32_t i = 0; i < symtab->sheafs.current_index; i++){
+		symtab_variable_sheaf_t* sheaf = dynamic_array_get_at(&(symtab->sheafs), i);
+
+		for(int32_t i = 0; i < VARIABLE_KEYSPACE; i++){
+			symtab_variable_record_t* cursor = sheaf->records[i];
+
+			while(cursor != NULL){
+
+				//TODO DEPENDENCY NODE
+				//TODO
+
+
+				cursor = cursor->next;
+			}
+		}
+	}
+}
+
+
 
 /**
  * Perform all static analysis on a given CFG. The functions
@@ -873,6 +1228,10 @@ cfg_construction_result_type_t perform_all_static_analysis(cfg_t* cfg, front_end
 	instruction_pointer_var = cfg->instruction_pointer;
 	stack_pointer_var = cfg->stack_pointer;
 
+	//Cache these as well
+	error_count = &(results->num_errors);
+	warning_count = &(results->num_warnings);
+
 	/**
 	 * 1.) Mangle all static variable names with a unique number identifier at the very end
 	 * to avoid name collisions
@@ -884,6 +1243,24 @@ cfg_construction_result_type_t perform_all_static_analysis(cfg_t* cfg, front_end
 	 * basis for all of our future checks & optimizations
 	 */
 	convert_cfg_to_ssa_form(cfg, results->variable_symtab);
+
+	/**
+	 * Step 3: perform definite assignment analysis on the entire
+	 * CFG. This process will verify that all variables are only
+	 * used after they are guaranteed to have been assigned
+	 *
+	 * NOTE: this is a potential fail point for the CFG
+	 */
+	if(perform_definite_assignment_analysis(cfg, results->variable_symtab) == FAILURE){
+		return CFG_RESULT_FAILURE;
+	}
+
+	/**
+	 * 4.) Perform mutability checking. Unlike definite assignment
+	 * analysis there is no chance for failure here, this
+	 * just generates warnings
+	 */
+	perform_mutability_checking(cfg, results->variable_symtab);
 
 	//TODO DUMMY FOR NOW
 	return CFG_RESULT_SUCCESS;
